weeeeee

+++ cell04/ex02/calculator.py
@@ -1,7 +1,7 @@
 print("Give me the first number: ")
-num1 = int(input())
+num1 = float(input())
 print("Give me the second number: ")
-num2 = int(input())
+num2 = float(input())
 print("Thank You!")
 try:
     print(f"{num1} + {num2} = {num1 + num2}")
@@ -11,4 +11,4 @@ try:
 
 except ZeroDivisionError:
     print(f"{num1} / {num2} = Undefined") # / 0 is not allowed & fix ZeroDivisionError error
-    print(f"{num1} x {num2} = {num1 * num2}") # add ts again to keep running till the end ts pmo :broken_heart:
+    print(f"{num1} x {num2} = {num1 * num2}") # add ts again to keep running till the end ts pmo :broken_heart:6
